requirements.txt
Flask==2.3.2
PyMuPDF==1.26.4
python-docx==0.8.11
scikit-learn==1.3.2
Werkzeug==2.3.7
